daemon: move infraendpoints from `daemon/cmd` to `daemon/infraendpoints`
In preparation to initialize the Cilium Health via Hive Hooks (and depend
on the `InfraIPAllocator` (for receiving the health routing info), this commit
moves the `InfraIPAllocator` from package `daemon/cmd` to `daemon/infraendpoints`
and introduces a Hive Cell & interface `InfraIPAllocator`.

Upcoming commits will also move the creation of the health & ingress endpoints
to this package & cell.

Note: Yes, naming is hard. I'm not really happy with `daemon/infraendpoints` -
but it's better than being in `daemon/cmd` and the logic is also not core IPAM
or endpoint (once we move the health & ingress endpoint creation into this package &
cell too.)
