+++ daemon/cmd/bootstrap_statistics.go
@@ -15,7 +15,6 @@ type bootstrapStatistics struct {
 	k8sInit     spanstat.SpanStat
 	restore     spanstat.SpanStat
 	healthCheck spanstat.SpanStat
-	ingressIPAM spanstat.SpanStat
 	cleanup     spanstat.SpanStat
 	bpfBase     spanstat.SpanStat
 	ipam        spanstat.SpanStat
@@ -48,7 +47,6 @@ func (b *bootstrapStatistics) getMap() map[string]*spanstat.SpanStat {
 		"k8sInit":     &b.k8sInit,
 		"restore":     &b.restore,
 		"healthCheck": &b.healthCheck,
-		"ingressIPAM": &b.ingressIPAM,
 		"cleanup":     &b.cleanup,
 		"bpfBase":     &b.bpfBase,
 		"ipam":        &b.ipam,
