+++ daemon/infraendpoints/infra_ip_allocation_test.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package cmd
+package infraendpoints
 
 import (
 	"fmt"
